pocld: move PoclDaemon to its own files
